Remove abstract
